refactor: extract dom snapshot runtime
